requirements.txt
anyio==3.6.1
asgiref==3.5.1
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
fastapi==0.78.0
h11==0.13.0
idna==3.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.4.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
